moves all web routes under web folder
