+++ app/Http/Controllers/Web/PostController.php
@@ -1,12 +1,13 @@
 <?php
 
-namespace Rogue\Http\Controllers;
+namespace Rogue\Http\Controllers\Web;
 
 use Rogue\Services\Fastly;
 use Rogue\Services\PostService;
 use Rogue\Repositories\SignupRepository;
 use Rogue\Http\Transformers\PostTransformer;
 use Rogue\Http\Controllers\Traits\PostRequests;
+use Rogue\Http\Controllers\Controller;
 
 class PostController extends Controller
 {
