+++ app/Http/Controllers/Web/ReviewsController.php
@@ -1,12 +1,13 @@
 <?php
 
-namespace Rogue\Http\Controllers;
+namespace Rogue\Http\Controllers\Web;
 
 use Rogue\Models\Post;
 use Rogue\Services\PostService;
 use Rogue\Http\Requests\ReviewsRequest;
 use Rogue\Http\Transformers\PostTransformer;
 use Illuminate\Database\Eloquent\ModelNotFoundException;
+use Rogue\Http\Controllers\Controller;
 
 class ReviewsController extends Controller
 {
